[refactoring] Separate sign-up State models, reorganize image resizing logic, and etc
- Split the State models for sign-up steps and email/nickname duplication checks
- Extracted a common BottomSheetDialog composable function for image uploads
- Changed image resizing logic position after pressing the final sign-up button
- Switched image resizing to use Dispatchers.Default
- Added actions for timeout
- Separated content views
- Moved format-check variables to an external location
- Replaced the existing Surface with a Scaffold
